agixt/requirements.txt
#### This block includes deps for local providers - all may big dependencies due to torch
--extra-index-url https://download.pytorch.org/whl/cpu torch
chromadb
git+https://github.com/huggingface/transformers
sentence-transformers
llama-cpp-python
accelerate
nomic
####
argparse
anthropic
beautifulsoup4
captcha-solver
docker
duckduckgo_search==2.9.5
fastapi
uvicorn
gitpython
google-api-python-client
GoogleBard
google-auth
google-auth-httplib2
google-auth-oauthlib
gtts
InstructorEmbedding
jsonschema
openai
playwright
playsound==1.2.2
pre-commit
pytz
pynacl
python-git
PyGithub
selenium
sendgrid
tiktoken
tweepy
webdriver_manager
spacy
discord
git+https://github.com/xtekky/gpt4free

-r https://raw.githubusercontent.com/xtekky/gpt4free/main/requirements.txt

accelerate
torch
EdgeGPT
google-generativeai
google-cloud-aiplatform
streamlit
docx2txt
pdfplumber
ffmpeg
hugchat
flask-bcrypt
revChatGPT
